Rename OrderSigningApi
