+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './common'
 export * from './order-book'
 export * from './subgraph'
-export * from './order-sign'
+export * from './order-signing'
