+++ src/order-signing/api.spec.ts
@@ -1,14 +1,14 @@
-import { OrderSignApi } from './api'
+import { OrderSigningApi } from './api'
 import { SupportedChainId } from '../common'
 import { UnsignedOrder } from './types'
 import { OrderKind } from '../order-book'
 import { mock, instance, when, anything } from 'ts-mockito'
 import type { TypedDataSigner } from '@cowprotocol/contracts'
 
-describe('OrderSignApi', () => {
+describe('OrderSigningApi', () => {
   const signature =
     '0xbb6aaa4207e9e97275934e26fc2fce3ed40a6d6b607870c1913cdb86b13070ec6196010f40e8ddfd36ad7f9eae033dfd5d4bf551feaf664a6ca768edde2d0a701c'
-  const orderSignApi = new OrderSignApi()
+  const orderSignApi = new OrderSigningApi()
   let signer: TypedDataSigner
 
   beforeEach(() => {
